refactor(voxedge): consolidate trt/ into jetson/ (single voxedge[jetson])
trt_edge_llm is a Jetson backend (registry jetson.trt_edge_llm); the separate
voxedge/backends/trt/ + voxedge[trt] extra was an artifact of incremental
migration sequencing, not a real boundary. Move trt_edge_llm_asr/tts/ipc +
worker_io + _trt_edge_llm_util into voxedge/backends/jetson/; merge the trt
extra deps (webrtcvad) into voxedge[jetson]; drop the trt package/extra. All
jetson.* backends now live under one dir + one extra, matching the registry.
Pure reorg, no behavior change. 27 passed.
